+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD
 
+* `sequel` – Rename `:callbacks` option to `:hooks` (@janko)
+
 * `derivatives` – Auto-download `UploadedFile` objects passed to `Attacher#process_derivatives` (@janko)
 
 ## 3.0.0.beta (2019-08-29)